MAINT: Extract the crazy number of type aliases to their own file
This is a straight copy-paste, followed by a docstring and some imports.
Note that we have to leave all the old imports in place to preserve the public API, even though they are no longer used.
